% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Conversion}
\usage{
convert(object, to)
}
\arguments{
\item{object}{an object of class \code{jointLPM}}

\item{to}{character. Either "multlcmm" or "Jointlcmm", indicating to which type the object should be converted.}
}
\value{
an object of class \code{multlcmm} or \code{Jointlcmm}.
}
\description{
This function converts a jointLPM object to an object of type \code{multlcmm} or
\code{Jointlcmm} from lcmm package.
The conversion to \code{multlcmm} unable the use of the dedicated postfit and
predictions functions implemented in the lcmm package (plot, predictL, predictY,
predictlink, predictYcond, fitY, VarExpl).
The conversion to \code{Jointlcmm} permits the use of functions cuminc and
plot (with options "baselinerisk" or "survival").
}
