% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRE.R
\name{createFargs}
\alias{createFargs}
\title{Likelihood arguments}
\usage{
createFargs(x, data)
}
\arguments{
\item{x}{a \code{jointLPM} object}

\item{data}{a data.frame containing the estimation dataset}
}
\value{
a named list
}
\description{
Create the arguments to be passed to the \code{loglik} function
}
