% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createX0.R
\name{createX0}
\alias{createX0}
\title{createX0}
\usage{
createX0(form, data)
}
\arguments{
\item{form}{a list of formula}

\item{data}{a data frame containing all variables mentionned in form}
}
\value{
a list containing
  \item{X0}{the pooled (with cbind) model matrix. The number of lines is the same as in 'data'}
 \item{idform}{a list of the same length as 'form' specifying which term of X0 appear in each formula}
}
\description{
Create the matrix containing all covariates appearing in the formula
}
\details{
Caution 1 : 'data' should not contain any missing values. There is no check for that within the function.
Caution 2 : if interaction terms are present, the names are sorted. For example x:age becomes age:x in X0.
Caution 3 : intercept is not systematically in the first column. It will only if form[[1]] contains an intercept.
}
