% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSojournTime.R
\name{diffSojournTime}
\alias{diffSojournTime}
\title{Difference of sojourn times}
\usage{
diffSojournTime(x1, x2)
}
\arguments{
\item{x1}{a sojournTime object, called with \code{draws = TRUE} and
\code{returndraws = TRUE}, computed for a specified seed.}

\item{x2}{a sojournTime object, called with \code{draws = TRUE} and
\code{returndraws = TRUE}, computed with the same seed as \code{x1}.}
}
\value{
returns the median, the 2.5\% and 97.5\% quantiles, the mean,
the standard deviation of the difference between the two sojourn times
and the number of removed draws (eventually due to computational issues)
}
\description{
This function computes the difference between two sojourn times.
}
