% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT4Sselecting.R
\name{IRT4Sselecting}
\alias{IRT4Sselecting}
\title{Fisher information function, used in Step 4 _ Selecting from the 4Smethod}
\usage{
IRT4Sselecting(D, proj)
}
\arguments{
\item{D}{an object of class \code{jointLPM} representing the estimated 
dimension model with the items}

\item{proj}{a vector}
}
\value{
a list of two matrices

###@examples
}
\description{
This function computes the contribution if each item during stages, based on 
the Fisher information.
}
\references{
preprint :  arXiv:2407.08278
}
\author{
Tiphaine Saulnier, Viviane Philipps and Cecile Proust-Lima
}
