% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLPM-package.R
\docType{package}
\name{JLPM-package}
\alias{JLPM-package}
\title{Joint latent process models}
\description{
Functions for the estimation of joint latent process models (JLPM)
}
\references{
Saulnier, Philipps, Meissner, Rascol, Pavy-Le-Traon, Foubert-Samier, Proust-Lima (2021).
Joint models for the longitudinal analysis of measurement scales in the presence 
of informative dropout   arXiv:2110.02612

Philipps, Hejblum, Prague, Commenges, Proust-Lima (2021).
Robust and efficient optimization using a Marquardt-Levenberg algorithm with 
R package marqLevAlg   arXiv:2009.03840  #TS
}
\author{
Cecile Proust-Lima, Viviane Philipps, Tiphaine Saulnier

\email{cecile.proust-lima@inserm.fr}
}
\keyword{package}
