% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointLPM.R
\name{loglik}
\alias{loglik}
\title{Wraper to the Fortran subroutine that computed the log-likelihood}
\usage{
loglik(
  b0,
  Y0,
  X0,
  Tentr0,
  Tevt0,
  Devt0,
  ind_survint0,
  idea0,
  idg0,
  idcor0,
  idcontr0,
  idsurv0,
  idtdv0,
  typrisq0,
  nz0,
  zi0,
  nbevt0,
  idtrunc0,
  logspecif0,
  ny0,
  ns0,
  nv0,
  nobs0,
  nmes0,
  idiag0,
  ncor0,
  nalea0,
  npm0,
  nfix0,
  bfix0,
  epsY0,
  idlink0,
  nbzitr0,
  zitr0,
  uniqueY0,
  indiceY0,
  nvalSPLORD0,
  fix0,
  methInteg0,
  nMC0,
  dimMC0,
  seqMC0,
  idst0,
  nXcl0,
  Xcl_Ti0,
  Xcl_GK0,
  expectancy0
)
}
\arguments{
\item{b0}{b0}

\item{Y0}{Y0}

\item{X0}{X0}

\item{Tentr0}{Tentr0}

\item{Tevt0}{Tevt0}

\item{Devt0}{Devt0}

\item{ind_survint0}{ind_survint0}

\item{idea0}{idea0}

\item{idg0}{idg0}

\item{idcor0}{idcor0}

\item{idcontr0}{idcontr0}

\item{idsurv0}{idsurv0}

\item{idtdv0}{idtdv0}

\item{typrisq0}{typrisq0}

\item{nz0}{nz0}

\item{zi0}{zi0}

\item{nbevt0}{nbevt0}

\item{idtrunc0}{idtrunc0}

\item{logspecif0}{logspecif0}

\item{ny0}{ny0}

\item{ns0}{ns0}

\item{nv0}{nv0}

\item{nobs0}{nobs0}

\item{nmes0}{nmes0}

\item{idiag0}{idiag0}

\item{ncor0}{ncor0}

\item{nalea0}{nalea0}

\item{npm0}{npm0}

\item{nfix0}{nfix0}

\item{bfix0}{bfix0}

\item{epsY0}{epsY0}

\item{idlink0}{idlink0}

\item{nbzitr0}{nbzitr0}

\item{zitr0}{zitr0}

\item{uniqueY0}{uniqueY0}

\item{indiceY0}{indiceY0}

\item{nvalSPLORD0}{nvalSPLORD0}

\item{fix0}{fix0}

\item{methInteg0}{methInteg0}

\item{nMC0}{nMC0}

\item{dimMC0}{dimMC0}

\item{seqMC0}{seqMC0}

\item{idst0}{idst0}

\item{nXcl0}{nXcl0}

\item{Xcl_Ti0}{Xcl_Ti0}

\item{Xcl_GK0}{Xcl_GK0}

\item{expectancy0}{expectancy0}
}
\value{
the log-likelihood
}
\description{
Computes the log-likelihood of a jointLPM model
}
