% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRE.R
\name{posteriorLogLikelihood}
\alias{posteriorLogLikelihood}
\title{Posterior density of the random effects for 1 subject}
\usage{
posteriorLogLikelihood(value, x, data)
}
\arguments{
\item{value}{numeric vector containing the value at which the log-likelihood will be computed}

\item{x}{a \code{jointLPM} object}

\item{data}{a data.frame object containing the observations of a single subject}
}
\value{
the conditional log-likelihood of the random effects given the observed data
}
\description{
Posterior density of the random effects for 1 subject
}
