% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRE.R
\name{predictRE}
\alias{predictRE}
\title{Random effects prediction}
\usage{
predictRE(x, data, control = NULL, variance = FALSE)
}
\arguments{
\item{x}{a \code{jointLPM} object}

\item{data}{a data.frame object containing the data from the predictions are
to be computed}

\item{control}{optional list of arguments to be passed to the
\code{marqLevAlg} optimization function.}

\item{variance}{logical indicating whether the variance of the estimated
predictions should be returned. Default ot FALSE.}
}
\value{
a matrix containing the predicted random effects (in columns) for
each subject (in rows). If \code{variance = TRUE}, returns a list containing
the previous matrix and a vector containing the upper part of each variance
matrix.
}
\description{
The random effects of a \code{jointLPM} model are predicted as the mode of 
the conditional distribution of the random effects given the subject's
longitudinal and survival observations.
}
