% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardMethods.R
\name{StandardMethods}
\alias{StandardMethods}
\alias{coef.jointLPM}
\alias{vcov.jointLPM}
\title{Standard methods for estimated models}
\usage{
\method{coef}{jointLPM}(object, ...)
}
\arguments{
\item{object}{an object of class \code{jointLPM}}

\item{...}{other arguments. There are ignored in these functions.}
}
\value{
For \code{coef}, the vector of the estimated parameters.

For \code{vcov}, the variance-covariance matrix of the estimated parameters.
}
\description{
coef and vcov  methods for estimated jointLPM models.
}
\author{
Viviane Philipps
}
