% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jointLPM.R
\name{summary.jointLPM}
\alias{summary.jointLPM}
\title{Summary of a \code{JLPM} object}
\usage{
\method{summary}{jointLPM}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{JLPM} for a joint latent process model.}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
Print a summary of JLPM estimations 
with matrices containing the estimates, their standard errors, Wald statistics and p-values
for the parameter of each submodel. #TS
}
\description{
This function provides a summary of \code{JLPM} estimations.
}
\author{
Viviane Philipps, Tiphaine Saulnier and Cecile Proust-Lima
}
