% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT4Sselecting.R
\name{IRT4Sselecting}
\alias{IRT4Sselecting}
\title{Fisher information function, used in Step 4 _ Selecting from the 4Smethod}
\usage{
IRT4Sselecting(D, proj)
}
\arguments{
\item{D}{an object of class \code{jointLPM} representing the estimated 
dimension model with the items}

\item{Dss}{an object of class \code{jointLPM} representing the estimated proxy 
model between the dimension scores and the stages}

\item{nsim}{number of points used in the numerical integration (Monte-Carlo). 
nsim should be relatively important (nsim=1000 by default).}

\item{bounds}{dimension scale boundaries between which are computed a grid of
the dum of the items}
}
\value{
a vector of the estimated stage transition thresholds

###@examples
}
\description{
This function computes the contribution if each item during stages, based on 
the Fisher information. 
Specifically, this function xxx. 
The output is a list of two matrices : xxx.
}
\references{
preprint :  arXiv:2407.08278
}
\author{
Tiphaine Saulnier, Viviane Philipps and Cecile Proust-Lima
}
