% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLPM-package.R
\docType{package}
\name{JLPM-package}
\alias{JLPM-package}
\title{Joint latent process models}
\description{
Functions for the estimation of joint latent process models (JLPM)
}
\references{

}
\author{
Cecile Proust-Lima, Viviane Philipps, Tiphaine Saulnier

\email{cecile.proust-lima@inserm.fr}
}
\keyword{package}
