% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jointLPM.R
\name{summary.jointLPM}
\alias{summary.jointLPM}
\title{Summary of a joint latent process model}
\usage{
\method{summary}{jointLPM}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{jointLPM} for a joint latent process model.}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
The function is mainly used for its side effects. It returns invisibily a list of two matrices
containing the estimates, their standard errors, Wald statistics and associated p-values for the survival
submodel (first element of the list) and for the mixed model's fixed effects (second element).
}
\description{
This function provides a summary of model estimated with the \code{jointLPM} function.
}
\author{
Viviane Philipps, Tiphaine Saulnier and Cecile Proust-Lima
}
